chore(tree): move sample data files to `app/data`
