+++ live-editing/configs/TreeConfigGenerator.ts
@@ -11,7 +11,7 @@ export class TreeConfigGenerator implements IConfigGenerator {
 
         configs.push(new Config({
             component: 'TreeBasicSampleComponent',
-            additionalFiles: ['/src/app/lists/tree/tree-basic-sample/local-data.ts'],
+            additionalFiles: ['/src/app/data/animations-data.ts'],
             appModuleConfig: new AppModuleConfig({
                 imports: ['IgxTreeModule', 'IgxIconModule', 'TreeBasicSampleComponent'],
                 ngDeclarations: ['TreeBasicSampleComponent'],
@@ -22,7 +22,7 @@ export class TreeConfigGenerator implements IConfigGenerator {
 
         configs.push(new Config({
             component: 'TreeAdvancedSampleComponent',
-            additionalFiles: ['/src/app/lists/tree/tree-advanced-sample/local-data.ts',
+            additionalFiles: ['/src/app/data/tree-file-data.ts',
             '/src/app/lists/tree/tree-advanced-sample/services/data.service.ts',
             '/src/app/lists/tree/tree-advanced-sample/services/svgIcons.ts'],
             appModuleConfig: new AppModuleConfig({
