+++ src/app/data/animations-data.ts
